# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ACLSubjectType:
    enum:
    - user
    - public
    type: string
    x-enum-varnames:
    - ACLSubjectTypeUser
    - ACLSubjectTypePublic
...
  handler.createACLRequest:
    properties:
      inherit:
        type: boolean
      permissionBit:
        type: integer
      subjectId:
        type: string
      subjectType:
        $ref: '#/definitions/domain.ACLSubjectType'
    required:
    - permissionBit
    - subjectType
    type: object
...
  handler.updateACLRequest:
    properties:
      inherit:
        type: boolean
      permissionBit:
        type: integer
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 科研项目文档管理系统后端 API。
  title: Research Admin Backend API
  version: "1.0"
paths:
...
  /documents/{documentId}/acl:
    get:
      description: 返回文档上显式配置的 ACL 规则，不包含 workspace 默认权限。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取文档 ACL
      tags:
      - document-acl
    post:
      consumes:
      - application/json
      description: 为文档新增一条 ACL 规则。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      - description: ACL 参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createACLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 新增文档 ACL
      tags:
      - document-acl
  /documents/{documentId}/acl/{aclId}:
    delete:
      description: 删除文档上的指定 ACL 规则。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      - description: ACL ID
        in: path
        name: aclId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除文档 ACL
      tags:
      - document-acl
    patch:
      consumes:
      - application/json
      description: 更新文档上的指定 ACL 规则。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      - description: ACL ID
        in: path
        name: aclId
        required: true
        type: string
      - description: ACL 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateACLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新文档 ACL
      tags:
      - document-acl
...
  /documents/{documentId}/my-permission:
    get:
      description: 返回当前用户对文档的最终权限位。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取我的文档权限
      tags:
      - document-acl
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"